test(web): move the MoneyGram prefill builders out of the component file
React Doctor flags non-component exports in a component file because Fast
Refresh cannot preserve state across them. The builders and their prefill
type now live in moneygram-prefill.ts; the widget imports them and its own
exports are the component and its props again.
